+++ foo_ui_columns/item_properties_config_list_view.cpp
@@ -17,7 +17,7 @@ void FieldsList::notify_save_inline_edit(const char* value)
     m_edit_index = pfc_infinite;
 }
 
-bool FieldsList::notify_create_inline_edit(const pfc::list_base_const_t<t_size>& indices, unsigned column,
+bool FieldsList::notify_create_inline_edit(const pfc::list_base_const_t<t_size>& indices, size_t column,
     pfc::string_base& p_text, t_size& p_flags, mmh::ComPtr<IUnknown>& pAutocompleteEntries)
 {
     t_size indices_count = indices.get_count();
@@ -36,7 +36,7 @@ bool FieldsList::notify_create_inline_edit(const pfc::list_base_const_t<t_size>&
 }
 
 bool FieldsList::notify_before_create_inline_edit(
-    const pfc::list_base_const_t<t_size>& indices, unsigned column, bool b_source_mouse)
+    const pfc::list_base_const_t<t_size>& indices, size_t column, bool b_source_mouse)
 {
     return column <= 1 && indices.get_count() == 1;
 }
