+++ foo_ui_columns/playlist_tabs.h
@@ -98,32 +98,30 @@ public:
     };
 
     void FB2KAPI on_items_removing(
-        unsigned p_playlist, const bit_array& p_mask, unsigned p_old_count, unsigned p_new_count) override;
+        size_t p_playlist, const bit_array& p_mask, size_t p_old_count, size_t p_new_count) override;
     // called before actually removing them
     void FB2KAPI on_items_removed(
-        unsigned p_playlist, const bit_array& p_mask, unsigned p_old_count, unsigned p_new_count) override;
+        size_t p_playlist, const bit_array& p_mask, size_t p_old_count, size_t p_new_count) override;
 
-    void on_playlist_activate(unsigned p_old, unsigned p_new) override;
+    void on_playlist_activate(size_t p_old, size_t p_new) override;
 
-    void on_playlists_reorder(const unsigned* p_order, unsigned p_count) override;
-    void on_playlist_created(unsigned p_index, const char* p_name, unsigned p_name_len) override;
-    void on_playlists_removed(const bit_array& p_mask, unsigned p_old_count, unsigned p_new_count) override;
-    void on_playlist_renamed(unsigned p_index, const char* p_new_name, unsigned p_new_name_len) override;
+    void on_playlists_reorder(const size_t* p_order, size_t p_count) override;
+    void on_playlist_created(size_t p_index, const char* p_name, size_t p_name_len) override;
+    void on_playlists_removed(const bit_array& p_mask, size_t p_old_count, size_t p_new_count) override;
+    void on_playlist_renamed(size_t p_index, const char* p_new_name, size_t p_new_name_len) override;
 
-    void on_items_added(
-        unsigned int, unsigned int, const pfc::list_base_const_t<metadb_handle_ptr>&, const bit_array&) override;
-    void on_items_reordered(unsigned int, const unsigned int*, unsigned int) override;
-    void on_items_selection_change(unsigned int, const bit_array&, const bit_array&) override;
-    void on_item_focus_change(unsigned int, unsigned int, unsigned int) override;
-    void on_items_modified(unsigned int, const bit_array&) override;
-    void on_items_modified_fromplayback(unsigned int, const bit_array&, play_control::t_display_level) override;
-    void on_items_replaced(
-        unsigned int, const bit_array&, const pfc::list_base_const_t<t_on_items_replaced_entry>&) override;
-    void on_item_ensure_visible(unsigned int, unsigned int) override;
-    void on_playlists_removing(const bit_array&, unsigned int, unsigned int) override;
+    void on_items_added(size_t, size_t, const pfc::list_base_const_t<metadb_handle_ptr>&, const bit_array&) override;
+    void on_items_reordered(size_t, const size_t*, size_t) override;
+    void on_items_selection_change(size_t, const bit_array&, const bit_array&) override;
+    void on_item_focus_change(size_t, size_t, size_t) override;
+    void on_items_modified(size_t, const bit_array&) override;
+    void on_items_modified_fromplayback(size_t, const bit_array&, play_control::t_display_level) override;
+    void on_items_replaced(size_t, const bit_array&, const pfc::list_base_const_t<t_on_items_replaced_entry>&) override;
+    void on_item_ensure_visible(size_t, size_t) override;
+    void on_playlists_removing(const bit_array&, size_t, size_t) override;
     void on_default_format_changed() override;
-    void on_playback_order_changed(unsigned int) override;
-    void on_playlist_locked(unsigned int, bool) override;
+    void on_playback_order_changed(size_t) override;
+    void on_playlist_locked(size_t, bool) override;
 
     void kill_switch_timer();
 
@@ -158,12 +156,12 @@ public:
     void get_supported_panels(
         const pfc::list_base_const_t<window::ptr>& p_windows, bit_array_var& p_mask_unsupported) override;
 
-    void insert_panel(unsigned index, const uie::splitter_item_t* p_item) override;
-    void remove_panel(unsigned index) override;
-    void replace_panel(unsigned index, const uie::splitter_item_t* p_item) override;
-    unsigned get_panel_count() const override;
-    unsigned get_maximum_panel_count() const override;
-    uie::splitter_item_t* get_panel(unsigned index) const override;
+    void insert_panel(size_t index, const uie::splitter_item_t* p_item) override;
+    void remove_panel(size_t index) override;
+    void replace_panel(size_t index, const uie::splitter_item_t* p_item) override;
+    size_t get_panel_count() const override;
+    size_t get_maximum_panel_count() const override;
+    uie::splitter_item_t* get_panel(size_t index) const override;
 
     void import_config(stream_reader* p_reader, t_size p_size, abort_callback& p_abort) override;
     void export_config(stream_writer* p_writer, abort_callback& p_abort) const override;
