+++ foo_ui_columns/splitter_panel.cpp
@@ -142,7 +142,7 @@ void FlatSplitterPanel::Panel::write(stream_writer* out, abort_callback& p_abort
     out->write_lendian_t(m_show_toggle_area, p_abort);
 
     refresh_child_data();
-    out->write_lendian_t(m_child_data.get_size(), p_abort);
+    out->write_lendian_t(gsl::narrow<uint32_t>(m_child_data.get_size()), p_abort);
     out->write(m_child_data.get_ptr(), m_child_data.get_size(), p_abort);
 
     out->write_lendian_t(m_use_custom_title, p_abort);
@@ -178,7 +178,7 @@ void FlatSplitterPanel::Panel::_export(stream_writer* out, abort_callback& p_abo
     out->write_lendian_t(m_autohide, p_abort);
     out->write_lendian_t(m_size.get_scaled_value(), p_abort);
     out->write_lendian_t(m_show_toggle_area, p_abort);
-    out->write_lendian_t(child_exported_data.m_data.get_size(), p_abort);
+    out->write_lendian_t(gsl::narrow<uint32_t>(child_exported_data.m_data.get_size()), p_abort);
     out->write(child_exported_data.m_data.get_ptr(), child_exported_data.m_data.get_size(), p_abort);
     out->write_lendian_t(m_use_custom_title, p_abort);
     out->write_string(m_custom_title, p_abort);
